+++ examples/ex32_harmonograph.clj
@@ -53,7 +53,7 @@
         a3 (r/drand)
         a4 (- 1.0 a3)
         n [(make-jnoise) (make-jnoise) j/perlin-noise j/perlin-noise r/noise r/noise r/noise]
-        palseq (filter #(> 240 ^double (c/to-luma (first %)) 100) (repeatedly #(:palette (g/color-reducer-machine))))
+        palseq (filter #(> 240 ^double (c/to-luma (first %)) 100) (repeatedly #(:palette (g/color-reducer-machine-random-config))))
         pal1 (first palseq)
         pal2 (second palseq)
         c1 (first pal1)